fix(storage): preserve cache correctness through compaction
