+++ crates/astrid-storage-content/src/read/traversal.rs
@@ -1,0 +1,875 @@
+use alloc::{sync::Arc, vec::Vec};
+
+use astrid_storage_model::{
+    ObjectClass, ObjectId, ObjectKind, ObjectRecord, ReferenceKind, ReferenceLabel,
+};
+
+use crate::{
+    CHUNK_TREE_FANOUT, CONTENT_LABEL, ChunkingProfile, ContentError, FORMAT_VERSION,
+    boundary::{is_canonical_boundary, is_canonical_final_chunk},
+    decode_file_header,
+};
+
+use super::{
+    ContentReadError, ContentSource, ContentVerificationDelta, ContentVerificationState,
+    VerifiedEdge,
+};
+
+#[derive(Clone, Copy)]
+pub(super) enum BoundaryMode<'a> {
+    Skip,
+    Validate,
+    Reuse(&'a ContentVerificationState),
+}
+
+pub(super) fn read_decoded_range<S: ContentSource>(
+    source: &S,
+    file: ObjectId,
+    decoded: (ChunkingProfile, u64, u64, Option<ObjectId>),
+    offset: u64,
+    length: u64,
+    boundary_mode: BoundaryMode<'_>,
+) -> Result<(Vec<u8>, ContentVerificationDelta), ContentReadError<S::Error>> {
+    let (profile, logical_bytes, chunk_count, content) = decoded;
+    let end = offset
+        .checked_add(length)
+        .ok_or(ContentError::RangeOutOfBounds {
+            offset,
+            length,
+            file_length: logical_bytes,
+        })?;
+    if offset > logical_bytes || end > logical_bytes {
+        return Err(ContentError::RangeOutOfBounds {
+            offset,
+            length,
+            file_length: logical_bytes,
+        }
+        .into());
+    }
+    let capacity = usize::try_from(length).map_err(|_| ContentError::LengthOverflow)?;
+    let mut output = Vec::new();
+    output
+        .try_reserve_exact(capacity)
+        .map_err(|_| ContentError::LengthOverflow)?;
+    if length == 0 {
+        return Ok((output, ContentVerificationDelta::default()));
+    }
+    let content = content.ok_or(ContentError::InvalidObject {
+        object: file,
+        detail: "non-empty file has no content reference",
+    })?;
+    let shape = ExpectedShape {
+        logical_bytes,
+        chunk_count,
+        tree_depth: canonical_tree_depth(chunk_count),
+        profile,
+        ends_file: true,
+    };
+    let mut boundaries = match boundary_mode {
+        BoundaryMode::Skip => None,
+        BoundaryMode::Validate => Some(BoundaryWindow::new(None)),
+        BoundaryMode::Reuse(known) => Some(BoundaryWindow::new(Some(known))),
+    };
+    let record = load(source, content)?;
+    let mut traversal = RangeTraversal {
+        range: RequestedRange { start: offset, end },
+        output: &mut output,
+        boundaries: &mut boundaries,
+    };
+    append_loaded_range(
+        source,
+        TraversalNode {
+            object: content,
+            record,
+            shape,
+            base_offset: 0,
+            path: Vec::new(),
+        },
+        &mut traversal,
+    )?;
+    if let Some(boundaries) = &mut boundaries {
+        boundaries.validate_neighbors(source, content, shape, logical_bytes)?;
+    }
+    if output.len() != capacity {
+        return Err(ContentError::InvalidObject {
+            object: file,
+            detail: "content tree reconstructed the wrong range length",
+        }
+        .into());
+    }
+    Ok((
+        output,
+        boundaries
+            .map(BoundaryWindow::into_delta)
+            .unwrap_or_default(),
+    ))
+}
+
+#[derive(Clone, Copy)]
+struct ExpectedShape {
+    logical_bytes: u64,
+    chunk_count: u64,
+    tree_depth: u32,
+    profile: ChunkingProfile,
+    ends_file: bool,
+}
+
+#[derive(Clone, Copy)]
+struct RequestedRange {
+    start: u64,
+    end: u64,
+}
+
+struct LoadedChunk {
+    object: ObjectId,
+    start: u64,
+    bytes: Vec<u8>,
+    path: Vec<TreeStep>,
+}
+
+struct TraversalNode {
+    object: ObjectId,
+    record: Arc<ObjectRecord>,
+    shape: ExpectedShape,
+    base_offset: u64,
+    path: Vec<TreeStep>,
+}
+
+struct RangeTraversal<'output, 'verification> {
+    range: RequestedRange,
+    output: &'output mut Vec<u8>,
+    boundaries: &'output mut Option<BoundaryWindow<'verification>>,
+}
+
+impl LoadedChunk {
+    fn end(&self) -> Result<u64, ContentError> {
+        self.start
+            .checked_add(u64::try_from(self.bytes.len()).map_err(|_| ContentError::LengthOverflow)?)
+            .ok_or(ContentError::LengthOverflow)
+    }
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+struct TreeStep {
+    tree: ObjectId,
+    child_index: u16,
+    child_count: u16,
+}
+
+struct BoundaryWindow<'a> {
+    known: Option<&'a ContentVerificationState>,
+    delta: ContentVerificationDelta,
+    first_start: Option<u64>,
+    first_prefix: Vec<u8>,
+    first_path: Vec<TreeStep>,
+    last: Option<LoadedChunk>,
+}
+
+impl<'a> BoundaryWindow<'a> {
+    fn new(known: Option<&'a ContentVerificationState>) -> Self {
+        Self {
+            known,
+            delta: ContentVerificationDelta::default(),
+            first_start: None,
+            first_prefix: Vec::new(),
+            first_path: Vec::new(),
+            last: None,
+        }
+    }
+
+    fn into_delta(self) -> ContentVerificationDelta {
+        self.delta
+    }
+
+    fn observe(
+        &mut self,
+        object: ObjectId,
+        start: u64,
+        bytes: &[u8],
+        path: &[TreeStep],
+        profile: ChunkingProfile,
+    ) -> Result<(), ContentError> {
+        if bytes.is_empty() {
+            return Err(ContentError::InvalidObject {
+                object,
+                detail: "content chunk is empty",
+            });
+        }
+        let prefix_length = bytes.len().min(2);
+        let prefix = &bytes[..prefix_length];
+        if let Some(previous) = self.last.take() {
+            if previous.end()? != start {
+                return Err(ContentError::InvalidObject {
+                    object,
+                    detail: "content traversal skipped an overlapping chunk",
+                });
+            }
+            let edge = edge_between(&previous.path, path, object, profile)?;
+            self.validate_edge(&previous, prefix, edge, profile)?;
+        } else {
+            self.first_start = Some(start);
+            self.first_prefix.extend_from_slice(prefix);
+            self.first_path.extend_from_slice(path);
+        }
+        self.last = Some(LoadedChunk {
+            object,
+            start,
+            bytes: bytes.to_vec(),
+            path: path.to_vec(),
+        });
+        Ok(())
+    }
+
+    fn validate_neighbors<S: ContentSource>(
+        &mut self,
+        source: &S,
+        content: ObjectId,
+        shape: ExpectedShape,
+        logical_bytes: u64,
+    ) -> Result<(), ContentReadError<S::Error>> {
+        let first_start = self.first_start.ok_or(ContentError::InvalidObject {
+            object: content,
+            detail: "non-empty range selected no content chunk",
+        })?;
+        if first_start > 0 {
+            let edge = edge_before(&self.first_path, content, shape.profile)?;
+            if !self.is_verified(edge) {
+                let previous = load_chunk_at_offset(
+                    source,
+                    content,
+                    shape,
+                    0,
+                    first_start.saturating_sub(1),
+                    Vec::new(),
+                )?;
+                let prefix = self.first_prefix.clone();
+                self.validate_edge(&previous, &prefix, edge, shape.profile)?;
+            }
+        }
+        let last = self.last.take().ok_or(ContentError::InvalidObject {
+            object: content,
+            detail: "non-empty range selected no final content chunk",
+        })?;
+        let last_end = last.end()?;
+        if last_end < logical_bytes {
+            let edge = edge_after(&last.path, content, shape.profile)?;
+            if !self.is_verified(edge) {
+                let next = load_chunk_at_offset(source, content, shape, 0, last_end, Vec::new())?;
+                let prefix_length = next.bytes.len().min(2);
+                self.validate_edge(&last, &next.bytes[..prefix_length], edge, shape.profile)?;
+            }
+        } else if shape.chunk_count > 1 && !is_canonical_final_chunk(&last.bytes, shape.profile) {
+            return Err(ContentError::InvalidObject {
+                object: last.object,
+                detail: "final chunk violates the declared FastCDC profile",
+            }
+            .into());
+        }
+        Ok(())
+    }
+
+    fn is_verified(&self, edge: VerifiedEdge) -> bool {
+        self.known.is_some_and(|known| known.contains(edge))
+            || self
+                .delta
+                .edges
+                .get(&edge.domain)
+                .is_some_and(|edges| edges & edge.mask() != 0)
+    }
+
+    fn validate_edge(
+        &mut self,
+        left: &LoadedChunk,
+        right_prefix: &[u8],
+        edge: VerifiedEdge,
+        profile: ChunkingProfile,
+    ) -> Result<(), ContentError> {
+        if self.is_verified(edge) {
+            return Ok(());
+        }
+        validate_boundary(left, right_prefix, profile)?;
+        self.delta.insert(edge);
+        Ok(())
+    }
+}
+
+fn edge_between(
+    left: &[TreeStep],
+    right: &[TreeStep],
+    object: ObjectId,
+    profile: ChunkingProfile,
+) -> Result<VerifiedEdge, ContentError> {
+    for (left, right) in left.iter().zip(right) {
+        if left == right {
+            continue;
+        }
+        if left.tree == right.tree && right.child_index == left.child_index.saturating_add(1) {
+            return Ok(VerifiedEdge::new(left.tree, left.child_index, profile));
+        }
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "adjacent chunks have inconsistent tree paths",
+        });
+    }
+    Err(ContentError::InvalidObject {
+        object,
+        detail: "adjacent chunks do not diverge at a tree edge",
+    })
+}
+
+fn edge_before(
+    path: &[TreeStep],
+    object: ObjectId,
+    profile: ChunkingProfile,
+) -> Result<VerifiedEdge, ContentError> {
+    path.iter()
+        .rev()
+        .find(|step| step.child_index > 0)
+        .map(|step| VerifiedEdge::new(step.tree, step.child_index.saturating_sub(1), profile))
+        .ok_or(ContentError::InvalidObject {
+            object,
+            detail: "non-initial chunk has no preceding tree edge",
+        })
+}
+
+fn edge_after(
+    path: &[TreeStep],
+    object: ObjectId,
+    profile: ChunkingProfile,
+) -> Result<VerifiedEdge, ContentError> {
+    path.iter()
+        .rev()
+        .find(|step| step.child_index.saturating_add(1) < step.child_count)
+        .map(|step| VerifiedEdge::new(step.tree, step.child_index, profile))
+        .ok_or(ContentError::InvalidObject {
+            object,
+            detail: "non-final chunk has no following tree edge",
+        })
+}
+
+fn validate_boundary(
+    left: &LoadedChunk,
+    right_prefix: &[u8],
+    profile: ChunkingProfile,
+) -> Result<(), ContentError> {
+    if is_canonical_boundary(&left.bytes, right_prefix, profile) {
+        Ok(())
+    } else {
+        Err(ContentError::InvalidObject {
+            object: left.object,
+            detail: "chunk boundary violates the declared FastCDC profile",
+        })
+    }
+}
+
+fn append_loaded_range<S: ContentSource>(
+    source: &S,
+    node: TraversalNode,
+    traversal: &mut RangeTraversal<'_, '_>,
+) -> Result<(), ContentReadError<S::Error>> {
+    match node.record.kind() {
+        ObjectKind::Chunk => append_chunk_range(node, traversal),
+        ObjectKind::ChunkTree => append_tree_range(source, node, traversal),
+        _ => Err(ContentError::InvalidObject {
+            object: node.object,
+            detail: "file content points to an unsupported object kind",
+        }
+        .into()),
+    }
+}
+
+fn append_chunk_range<E>(
+    node: TraversalNode,
+    traversal: &mut RangeTraversal<'_, '_>,
+) -> Result<(), ContentReadError<E>> {
+    let TraversalNode {
+        object,
+        record,
+        shape,
+        base_offset,
+        path,
+    } = node;
+    validate_chunk(object, &record, shape)?;
+    if let Some(boundaries) = traversal.boundaries.as_mut() {
+        boundaries.observe(
+            object,
+            base_offset,
+            record.canonical_bytes(),
+            &path,
+            shape.profile,
+        )?;
+    }
+    let chunk_end = base_offset
+        .checked_add(shape.logical_bytes)
+        .ok_or(ContentError::LengthOverflow)?;
+    let start = usize::try_from(
+        traversal
+            .range
+            .start
+            .max(base_offset)
+            .saturating_sub(base_offset),
+    )
+    .map_err(|_| ContentError::LengthOverflow)?;
+    let end = usize::try_from(
+        traversal
+            .range
+            .end
+            .min(chunk_end)
+            .saturating_sub(base_offset),
+    )
+    .map_err(|_| ContentError::LengthOverflow)?;
+    let bytes = record
+        .canonical_bytes()
+        .get(start..end)
+        .ok_or(ContentError::InvalidObject {
+            object,
+            detail: "chunk range exceeds payload",
+        })?;
+    traversal.output.extend_from_slice(bytes);
+    Ok(())
+}
+
+fn append_tree_range<S: ContentSource>(
+    source: &S,
+    node: TraversalNode,
+    traversal: &mut RangeTraversal<'_, '_>,
+) -> Result<(), ContentReadError<S::Error>> {
+    let TraversalNode {
+        object,
+        record,
+        shape,
+        base_offset,
+        path,
+    } = node;
+    if shape.tree_depth == 0 {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "chunk tree exceeds canonical depth",
+        }
+        .into());
+    }
+    let children = decode_tree(object, &record, shape)?;
+    let mut cursor = 0_u64;
+    let child_count = children.len();
+    let child_count_u16 = u16::try_from(child_count).map_err(|_| ContentError::LengthOverflow)?;
+    let mut selected = Vec::new();
+    for (index, (child, child_length, child_chunk_count)) in children.into_iter().enumerate() {
+        let child_end = cursor
+            .checked_add(child_length)
+            .ok_or(ContentError::LengthOverflow)?;
+        let child_base = base_offset
+            .checked_add(cursor)
+            .ok_or(ContentError::LengthOverflow)?;
+        let child_absolute_end = base_offset
+            .checked_add(child_end)
+            .ok_or(ContentError::LengthOverflow)?;
+        if traversal.range.start < child_absolute_end && traversal.range.end > child_base {
+            let child_index = u16::try_from(index).map_err(|_| ContentError::LengthOverflow)?;
+            let mut child_path = path.clone();
+            child_path.push(TreeStep {
+                tree: object,
+                child_index,
+                child_count: child_count_u16,
+            });
+            selected.push((
+                child,
+                ExpectedShape {
+                    logical_bytes: child_length,
+                    chunk_count: child_chunk_count,
+                    tree_depth: shape.tree_depth.saturating_sub(1),
+                    profile: shape.profile,
+                    ends_file: shape.ends_file && index.saturating_add(1) == child_count,
+                },
+                child_base,
+                child_path,
+            ));
+        }
+        cursor = child_end;
+    }
+    let ids: Vec<_> = selected.iter().map(|(child, _, _, _)| *child).collect();
+    let records = load_many(source, &ids)?;
+    for ((child, child_shape, child_base, child_path), record) in selected.into_iter().zip(records)
+    {
+        append_loaded_range(
+            source,
+            TraversalNode {
+                object: child,
+                record,
+                shape: child_shape,
+                base_offset: child_base,
+                path: child_path,
+            },
+            traversal,
+        )?;
+    }
+    Ok(())
+}
+
+fn load_many<S: ContentSource>(
+    source: &S,
+    objects: &[ObjectId],
+) -> Result<Vec<Arc<ObjectRecord>>, ContentReadError<S::Error>> {
+    let Some(first) = objects.first().copied() else {
+        return Ok(Vec::new());
+    };
+    let records = source
+        .load_shared_content_objects(objects)
+        .map_err(ContentReadError::Source)?;
+    if records.len() != objects.len() {
+        return Err(ContentError::InvalidObject {
+            object: first,
+            detail: "content source returned the wrong batch length",
+        }
+        .into());
+    }
+    objects
+        .iter()
+        .copied()
+        .zip(records)
+        .map(|(object, record)| {
+            record.ok_or_else(|| ContentReadError::Content(ContentError::MissingObject(object)))
+        })
+        .collect()
+}
+
+fn load_chunk_at_offset<S: ContentSource>(
+    source: &S,
+    object: ObjectId,
+    shape: ExpectedShape,
+    base_offset: u64,
+    target_offset: u64,
+    path: Vec<TreeStep>,
+) -> Result<LoadedChunk, ContentReadError<S::Error>> {
+    let record = load(source, object)?;
+    match record.kind() {
+        ObjectKind::Chunk => {
+            validate_chunk(object, &record, shape)?;
+            let end = base_offset
+                .checked_add(shape.logical_bytes)
+                .ok_or(ContentError::LengthOverflow)?;
+            if target_offset < base_offset || target_offset >= end {
+                return Err(ContentError::InvalidObject {
+                    object,
+                    detail: "chunk lookup offset is outside the payload",
+                }
+                .into());
+            }
+            Ok(LoadedChunk {
+                object,
+                start: base_offset,
+                bytes: record.canonical_bytes().to_vec(),
+                path,
+            })
+        },
+        ObjectKind::ChunkTree => {
+            if shape.tree_depth == 0 {
+                return Err(ContentError::InvalidObject {
+                    object,
+                    detail: "chunk tree exceeds canonical depth",
+                }
+                .into());
+            }
+            let children = decode_tree(object, &record, shape)?;
+            let mut cursor = 0_u64;
+            let child_count = children.len();
+            for (index, (child, child_length, child_chunk_count)) in
+                children.into_iter().enumerate()
+            {
+                let child_base = base_offset
+                    .checked_add(cursor)
+                    .ok_or(ContentError::LengthOverflow)?;
+                let child_end = child_base
+                    .checked_add(child_length)
+                    .ok_or(ContentError::LengthOverflow)?;
+                if target_offset >= child_base && target_offset < child_end {
+                    let child_index =
+                        u16::try_from(index).map_err(|_| ContentError::LengthOverflow)?;
+                    let child_count_u16 =
+                        u16::try_from(child_count).map_err(|_| ContentError::LengthOverflow)?;
+                    let mut child_path = path;
+                    child_path.push(TreeStep {
+                        tree: object,
+                        child_index,
+                        child_count: child_count_u16,
+                    });
+                    return load_chunk_at_offset(
+                        source,
+                        child,
+                        ExpectedShape {
+                            logical_bytes: child_length,
+                            chunk_count: child_chunk_count,
+                            tree_depth: shape.tree_depth.saturating_sub(1),
+                            profile: shape.profile,
+                            ends_file: shape.ends_file && index.saturating_add(1) == child_count,
+                        },
+                        child_base,
+                        target_offset,
+                        child_path,
+                    );
+                }
+                cursor = cursor
+                    .checked_add(child_length)
+                    .ok_or(ContentError::LengthOverflow)?;
+            }
+            Err(ContentError::InvalidObject {
+                object,
+                detail: "chunk lookup found no covering child",
+            }
+            .into())
+        },
+        _ => Err(ContentError::InvalidObject {
+            object,
+            detail: "file content points to an unsupported object kind",
+        }
+        .into()),
+    }
+}
+
+pub(super) fn decode_file(
+    object: ObjectId,
+    record: &ObjectRecord,
+) -> Result<(ChunkingProfile, u64, u64, Option<ObjectId>), ContentError> {
+    if record.kind() != ObjectKind::File
+        || record.format_version() != FORMAT_VERSION
+        || record.class() != ObjectClass::Metadata
+        || record.logical_bytes() != 0
+    {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "invalid file object type or accounting",
+        });
+    }
+    let (profile, logical_bytes, chunk_count) =
+        decode_file_header(object, record.canonical_bytes())?;
+    if logical_bytes != 0
+        && (chunk_count == 1) != (logical_bytes <= u64::from(profile.maximum_bytes()))
+    {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "file chunk count violates the whole-object threshold",
+        });
+    }
+    let label = ReferenceLabel::new(CONTENT_LABEL);
+    let content = record.reference(&label);
+    match (
+        logical_bytes,
+        chunk_count,
+        content,
+        record.references().len(),
+    ) {
+        (0, 0, None, 0) => Ok((profile, 0, 0, None)),
+        (0, _, _, _) => Err(ContentError::InvalidObject {
+            object,
+            detail: "empty file has chunks or content",
+        }),
+        (_, 0, _, _) => Err(ContentError::InvalidObject {
+            object,
+            detail: "non-empty file has no chunks",
+        }),
+        (_, _, Some(reference), 1) if reference.kind() == ReferenceKind::Owns => Ok((
+            profile,
+            logical_bytes,
+            chunk_count,
+            Some(reference.target()),
+        )),
+        _ => Err(ContentError::InvalidObject {
+            object,
+            detail: "invalid file content reference",
+        }),
+    }
+}
+
+fn validate_chunk(
+    object: ObjectId,
+    record: &ObjectRecord,
+    shape: ExpectedShape,
+) -> Result<(), ContentError> {
+    let actual =
+        u64::try_from(record.canonical_bytes().len()).map_err(|_| ContentError::LengthOverflow)?;
+    if record.format_version() != FORMAT_VERSION
+        || record.class() != ObjectClass::Data
+        || record.logical_bytes() != 0
+        || !record.references().is_empty()
+        || actual != shape.logical_bytes
+        || shape.chunk_count != 1
+        || shape.tree_depth != 0
+    {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "invalid chunk object",
+        });
+    }
+    validate_profile_bounds(
+        object,
+        actual,
+        shape.chunk_count,
+        shape.profile,
+        shape.ends_file,
+    )?;
+    Ok(())
+}
+
+fn decode_tree(
+    object: ObjectId,
+    record: &ObjectRecord,
+    shape: ExpectedShape,
+) -> Result<Vec<(ObjectId, u64, u64)>, ContentError> {
+    let bytes = record.canonical_bytes();
+    if record.format_version() != FORMAT_VERSION
+        || record.class() != ObjectClass::Metadata
+        || record.logical_bytes() != 0
+        || bytes.len() < 18
+    {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "invalid chunk-tree object",
+        });
+    }
+    let count = usize::from(u16::from_le_bytes(
+        bytes[0..2]
+            .try_into()
+            .map_err(|_| ContentError::LengthOverflow)?,
+    ));
+    let logical_bytes = u64::from_le_bytes(
+        bytes[2..10]
+            .try_into()
+            .map_err(|_| ContentError::LengthOverflow)?,
+    );
+    let chunk_count = u64::from_le_bytes(
+        bytes[10..18]
+            .try_into()
+            .map_err(|_| ContentError::LengthOverflow)?,
+    );
+    if count == 0
+        || count > CHUNK_TREE_FANOUT
+        || record.references().len() != count
+        || bytes.len() != 18_usize.saturating_add(count.saturating_mul(16))
+        || logical_bytes != shape.logical_bytes
+        || chunk_count != shape.chunk_count
+    {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "chunk-tree header is inconsistent",
+        });
+    }
+    let mut children = Vec::with_capacity(count);
+    let mut total = 0_u64;
+    let mut total_chunks = 0_u64;
+    let child_capacity = tree_capacity(shape.tree_depth.saturating_sub(1));
+    for (index, reference) in record.references().iter().enumerate() {
+        let expected_label = u16::try_from(index)
+            .map_err(|_| ContentError::LengthOverflow)?
+            .to_be_bytes();
+        if reference.label().as_bytes() != expected_label || reference.kind() != ReferenceKind::Owns
+        {
+            return Err(ContentError::InvalidObject {
+                object,
+                detail: "chunk-tree child reference is non-canonical",
+            });
+        }
+        let start = 18_usize.saturating_add(index.saturating_mul(16));
+        let child_length = u64::from_le_bytes(
+            bytes[start..start.saturating_add(8)]
+                .try_into()
+                .map_err(|_| ContentError::LengthOverflow)?,
+        );
+        let child_chunk_count = u64::from_le_bytes(
+            bytes[start.saturating_add(8)..start.saturating_add(16)]
+                .try_into()
+                .map_err(|_| ContentError::LengthOverflow)?,
+        );
+        if child_length == 0
+            || child_chunk_count == 0
+            || child_chunk_count > child_capacity
+            || (index.saturating_add(1) < count && child_chunk_count != child_capacity)
+        {
+            return Err(ContentError::InvalidObject {
+                object,
+                detail: "chunk-tree child shape is non-canonical",
+            });
+        }
+        validate_profile_bounds(
+            object,
+            child_length,
+            child_chunk_count,
+            shape.profile,
+            shape.ends_file && index.saturating_add(1) == count,
+        )?;
+        total = total
+            .checked_add(child_length)
+            .ok_or(ContentError::LengthOverflow)?;
+        total_chunks = total_chunks
+            .checked_add(child_chunk_count)
+            .ok_or(ContentError::LengthOverflow)?;
+        children.push((reference.target(), child_length, child_chunk_count));
+    }
+    if total != logical_bytes || total_chunks != chunk_count {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "chunk-tree child totals do not sum to header",
+        });
+    }
+    Ok(children)
+}
+
+fn validate_profile_bounds(
+    object: ObjectId,
+    logical_bytes: u64,
+    chunk_count: u64,
+    profile: ChunkingProfile,
+    ends_file: bool,
+) -> Result<(), ContentError> {
+    let maximum = u64::from(profile.maximum_bytes());
+    let maximum_total = chunk_count
+        .checked_mul(maximum)
+        .ok_or(ContentError::LengthOverflow)?;
+    let required_full_chunks = if ends_file {
+        chunk_count.saturating_sub(1)
+    } else {
+        chunk_count
+    };
+    // FastCDC 2020 begins its two-byte loop at floor(minimum / 2), so an odd
+    // declared minimum has an effective non-final lower bound one byte lower.
+    let effective_minimum = u64::from(profile.minimum_bytes() & !1);
+    let mut minimum_total = required_full_chunks
+        .checked_mul(effective_minimum)
+        .ok_or(ContentError::LengthOverflow)?;
+    if ends_file {
+        minimum_total = minimum_total
+            .checked_add(1)
+            .ok_or(ContentError::LengthOverflow)?;
+    }
+    if logical_bytes < minimum_total || logical_bytes > maximum_total {
+        return Err(ContentError::InvalidObject {
+            object,
+            detail: "content shape violates the declared chunking profile",
+        });
+    }
+    Ok(())
+}
+
+fn canonical_tree_depth(chunk_count: u64) -> u32 {
+    let mut depth = 0_u32;
+    let mut capacity = 1_u64;
+    while capacity < chunk_count {
+        capacity = capacity.saturating_mul(CHUNK_TREE_FANOUT as u64);
+        depth = depth.saturating_add(1);
+    }
+    depth
+}
+
+fn tree_capacity(depth: u32) -> u64 {
+    (0..depth).fold(1_u64, |capacity, _| {
+        capacity.saturating_mul(CHUNK_TREE_FANOUT as u64)
+    })
+}
+
+pub(super) fn load<S: ContentSource>(
+    source: &S,
+    object: ObjectId,
+) -> Result<Arc<ObjectRecord>, ContentReadError<S::Error>> {
+    source
+        .load_shared_content_object(object)
+        .map_err(ContentReadError::Source)?
+        .ok_or_else(|| ContentError::MissingObject(object).into())
+}
