+++ crates/astrid-storage-engine/src/durable/compaction/mod.rs
@@ -498,13 +498,8 @@ where
         let super::DurableInner { files, index, .. } = inner;
         let files = live_files_mut(files)?;
         for (id, location) in index.iter() {
-            let record = read_indexed_object(
-                &mut files.arena,
-                *id,
-                *location,
-                &self.identity,
-                self.limits,
-            )?;
+            let record =
+                read_indexed_object(&files.arena, *id, *location, &self.identity, self.limits)?;
             bytes.extend_from_slice(id.as_bytes());
             let count = u64::try_from(record.references().len())
                 .map_err(|_| DurableError::EncodingOverflow)?;
@@ -797,6 +792,8 @@ where
             .try_clone()
             .map_err(|source| io_error("clone compacted arena for positional reads", source))?;
         let arena_generation = inner.arena_generation.wrapping_add(1);
+        self.object_cache
+            .retain_objects(|object| replacement.index.contains_key(&object));
         inner.roots_by_principal = replacement.roots;
         inner.index = replacement.index;
         inner.pending_index_locations.clear();
@@ -831,7 +828,7 @@ where
                 .copied()
                 .ok_or(astrid_storage_model::ModelError::MissingObject(*id))?;
             let record =
-                read_indexed_object(&mut files.arena, *id, location, &self.identity, self.limits)?;
+                read_indexed_object(&files.arena, *id, location, &self.identity, self.limits)?;
             let payload = encode_object_frame(self.identity.scheme(), *id, &record)?;
             ensure_payload_limit(ARENA_FILE, 0, payload.len(), self.limits)?;
             let location = append_frame(&mut arena, ARENA_MAGIC, &payload)?;
